% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_functions_and_arguments.R
\name{assign_targets}
\alias{assign_targets}
\title{assign_targets}
\usage{
assign_targets(fn, env = globalenv())
}
\arguments{
\item{fn}{function name}

\item{env}{environment}
}
\description{
Load all argument defaults for a function for testing purposes
}
