% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{check_names}
\usage{
check_names(x, vars)
}
\arguments{
\item{x}{a data.frame, matrix, or list}

\item{vars}{column names to check}
}
\description{
Checks a data.frame, matrix, or list for a set of names. If they don't match, best matches are provided in a message to the user
}
