% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lines.R
\name{count_lines}
\alias{count_lines}
\title{count_lines}
\usage{
count_lines(dir, recursive = TRUE)
}
\arguments{
\item{dir}{path to a folder}

\item{recursive}{default to true}
}
\description{
counts lines of code in a directory
}
