% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{dropbox_root}
\alias{dropbox_root}
\title{dropbox_root}
\usage{
dropbox_root()
}
\description{
Finds the root of a dropbox path. Useful if you use dropbox on two different machines and you want your paths to work.
}
