% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_error.R
\name{explain_error}
\alias{explain_error}
\title{explain_error
Use the copilotcli to explain the previous error message}
\usage{
explain_error()
}
\description{
explain_error
Use the copilotcli to explain the previous error message
}
