% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_functions_and_arguments.R
\name{find_fn}
\alias{find_fn}
\title{find_fn}
\usage{
find_fn(fn, folder = getwd(), open = T)
}
\arguments{
\item{fn}{function name}

\item{folder}{folder name}

\item{open}{whether or not to try and open the file.}
}
\description{
find a function hiding in a folder of R docs
}
