% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_functions_and_arguments.R
\name{get_functions}
\alias{get_functions}
\title{get_functions}
\usage{
get_functions(path)
}
\arguments{
\item{path}{the filepath you wish to process}
}
\description{
gets functions from source file without running it.
}
