% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_functions_and_arguments.R
\name{load_dependencies}
\alias{load_dependencies}
\title{load_dependencies}
\usage{
load_dependencies(package)
}
\arguments{
\item{package}{a string}
}
\description{
loads all dependencies for a package
}
