% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign functions and arguments.R
\name{peruse_fns}
\alias{peruse_fns}
\title{peruse_fns}
\usage{
peruse_fns(package, cat = T)
}
\arguments{
\item{package}{name}

\item{cat}{concatenate and print output?}
}
\description{
Returns all functions from a package
}
