% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_to_cran.R
\name{reset_packages_to_cran}
\alias{reset_packages_to_cran}
\title{reset_packages_to_cran}
\usage{
reset_packages_to_cran()
}
\description{
finds version mismatches and updates back to current cran version
}
