% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_targets.R
\name{search_targets}
\alias{search_targets}
\title{search_targets}
\usage{
search_targets(pattern)
}
\arguments{
\item{pattern}{pattern to search for}
}
\description{
Search for target names within a manifest
}
