% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_functions_and_arguments.R
\name{unpack_fns}
\alias{unpack_fns}
\title{unpack_fns}
\usage{
unpack_fns(package, env = globalenv())
}
\arguments{
\item{package}{name}

\item{env}{environment}
}
\description{
Load all functions from a package into the global environment
}
