% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_functions_and_arguments.R
\name{update_git}
\alias{update_git}
\title{get_functions}
\usage{
update_git(m = "update", branch = NULL)
}
\arguments{
\item{m}{commit message}

\item{branch}{branch name}
}
\description{
gets functions from source file without running it.
}
