% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_openvscode.R
\name{write_opencode}
\alias{write_opencode}
\title{write_opencode}
\usage{
write_opencode(dir)
}
\arguments{
\item{dir}{path to a folder}
}
\description{
Write a file that opens vscode in the enclosing working directory
}
\details{
Only works for windows computers.
}
