% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_numeric.R
\name{safe_numeric}
\alias{safe_numeric}
\title{safe_numeric}
\usage{
safe_numeric(x, ignore = c())
}
\arguments{
\item{x}{A vector to convert to numeric}

\item{ignore}{A vector of values to ignore when checking for unknown fitted.values#' @return A numeric vector}
}
\description{
Convert a vector to numeric, but stop if unknown values could not be converted
}
